refactor(cli): promote host start/stop/status to top-level, rename auth status → whoami
Removes the confusing `host` (verb) vs `hosts` (noun) collision. `host`
held only start/stop/status — three local-service controls — while
`hosts` listed remote machines in the org. They sounded the same on the
command line and meant entirely different things.

- `superset host start`  → `superset start`
- `superset host stop`   → `superset stop`
- `superset host status` → `superset status`
- `superset auth status` → `superset auth whoami`

The `host` group is gone. `hosts` keeps its meaning (list remote hosts).

The auth rename follows npm/vercel/hcloud — `whoami` reads more
naturally for "who am I logged in as" than `status` does, and avoids
the implicit asymmetry of having `superset status` mean "host service
state" while `superset auth status` means "auth state". Identity now
gets its own verb.
