+++ packages/cli/src/commands/auth/whoami/command.ts
