--- packages/cli/src/commands/host/meta.ts
@@ -1,4 +1,0 @@
-export default {
-	description: "Manage host service",
-	standalone: true,
-};
